+++ src/test/scala/uk/gov/hmrc/test/api/models/sol/FCSolCalculationSummaryResponse.scala
@@ -8,7 +8,7 @@ package uk.gov.hmrc.test.api.models.sol
 import play.api.libs.json.{Json, OFormat}
 
 case class FCSolCalculationSummaryResponse(
-  amountIntTotal: Int,
+  amountIntTotal: BigDecimal,
   combinedDailyAccrual: Int,
   debts: List[FCSolCalculation]
 )
